requirements.txt
# Core Machine Learning & Data Science
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
numpy==1.26.4
pandas>=2.1.4
scipy==1.12.0
scikit-learn==1.6.1
statsmodels==0.14.4
rdkit==2024.9.5
ultralytics>=8.3.81

# Deep Learning & AI
transformers @ git+https://github.com/huggingface/transformers@dd16acb8a3e93b643aa374c9fb80749f5235c1a6
huggingface-hub==0.28.1
accelerate==1.3.0
peft==0.14.0
openai==1.63.0
google-generativeai>=0.3.0
sentence-transformers==2.5.1
safetensors==0.5.2
ollama==0.4.7
roman-numbers==1.0.2

# Computer Vision & Image Processing
pillow==11.1.0
opencv-python==4.11.0.86
imageio==2.37.0
scikit-image==0.24.0
Pillow>=9.5.0

# Natural Language Processing
spacy==3.8.3
nltk==3.9.1
gensim==4.3.3
tokenizers==0.21.0
fasttext==0.9.3
rank-bm25==0.2.2

# Document Processing
python-docx==1.1.2
openpyxl==3.1.5
PyMuPDF==1.25.3
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.5
pytesseract==0.3.13
camelot-py>=0.11.0
tabula-py>=2.9.0
ghostscript==0.7

# ONNX Runtime (Platform-specific)
onnxruntime==1.19.2
onnxruntime-gpu>=1.15.0; platform_system != "Darwin"
onnxruntime-silicon>=1.15.0; platform_system == "Darwin" and platform_machine == "arm64"

# API & Web
fastapi==0.115.8
starlette==0.45.3
aiohttp>=3.9.0
httpx==0.28.1
requests>=2.31.0

# Data Formats & Serialization
PyYAML==6.0.2
json-repair==0.36.1
protobuf==5.29.3
pyarrow==19.0.0

# Utilities
tqdm==4.67.1
pydantic==2.10.6
python-multipart==0.0.20
python-dotenv>=1.0.0
colorama==0.4.6
rich==13.9.4

# Development & Testing
black==25.1.0
flake8==7.1.1
mypy==1.15.0
pytest==8.3.4
coverage==7.6.12

# Visualization
matplotlib==3.9.4
seaborn==0.13.2
plotly==6.0.0

# Database
SQLAlchemy==2.0.38
psycopg2-binary==2.9.10
redis==5.2.1

# Additional Libraries
aiohappyeyeballs==2.4.6
aiohttp-cors==0.7.0
aioredis==2.0.1
aiosignal==1.3.2
alabaster==0.7.16
annotated-types==0.7.0
antlr4-python3-runtime==4.7.2
anyio==4.8.0
anytree==2.12.1
async-timeout==5.0.1
asyncio==3.4.3
attrs==25.1.0
audioread==3.0.1
autograd==1.7.0
av==14.1.0
babel==2.17.0
beartype==0.18.5
beautifulsoup4==4.13.3
blis==1.2.0
cachetools==5.5.1
catalogue==2.0.10
cbor==1.0.0
cbor2==5.6.5
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
cn2an==0.5.23
coloredlogs==15.0.1
colorful==0.5.6
confection==0.1.5
contourpy==1.3.0
cramjam==2.9.1
cryptography==44.0.1
cycler==0.12.1
cymem==2.0.11
Cython==3.0.12
datasets==3.3.0
datrie==0.8.2
decorator==5.1.1
dependency-injector==4.45.0
dill==0.3.8
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
docutils==0.21.2
editdistance-s==1.0.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
exceptiongroup==1.2.2
execnet==2.1.1
faiss-cpu==1.10.0
Faker==36.1.1
fastparquet==2024.11.0
ffmpeg-python==0.2.0
filelock==3.17.0
FlagEmbedding==1.3.4
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
ftfy==6.3.1
future==1.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.67.0
graspologic==3.4.1
graspologic-native==1.2.3
grpcio==1.70.0
grpcio-status==1.70.0
grpcio-tools==1.70.0
h11==0.14.0
h2==4.2.0
hanziconv==0.3.2
hpack==4.1.0
httpcore==1.0.7
httplib2==0.22.0
humanfriendly==10.0
hyperframe==6.1.0
HyperPyYAML==1.2.2
hypothesis==6.125.3
hyppo==0.4.0
idna==3.10
ijson==3.3.0
imagesize==1.4.1
importlib_metadata==8.6.1
importlib_resources==6.5.2
iniconfig==2.0.0
inscriptis==2.5.3
ir_datasets==0.5.9
isort==6.0.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keybert==0.9.0
kiwisolver==1.4.7
langcodes==3.5.0
langdetect==1.0.9
language_data==1.3.0
latex2sympy2==1.9.1
lazy_loader==0.4
librosa==0.10.2.post1
llama_cpp_python==0.3.7
llvmlite==0.43.0
lxml==5.3.1
lz4==4.4.3
marisa-trie==1.2.1
markdown-it-py==3.0.0
markdown_to_json==2.1.2
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.12
mypy-extensions==1.0.0
narwhals==1.26.0
networkx==3.2.1
numba==0.60.0
nvidia-ml-py==12.570.86
openai-whisper==20240930
opencensus==0.11.4
opencensus-context==0.1.3
orjson==3.10.15
packaging==24.2
pathspec==0.12.1
patsy==1.0.1
platformdirs==4.3.6
pluggy==1.5.0
pooch==1.8.2
portalocker==2.10.1
POT==0.9.5
preshed==3.0.9
proces==0.1.7
prometheus_client==0.21.1
propcache==0.2.1
proto-plus==1.26.0
psutil==7.0.0
py-cpuinfo==9.0.0
py-spy==0.4.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybind11==2.13.6
pyclipper==1.3.0.post6
pycodestyle==2.12.1
pycparser==2.22
pydantic_core==2.27.2
pydub==0.25.1
pyflakes==3.2.0
Pygments==2.19.1
pynndescent==0.5.13
pynvml==12.0.0
pyparsing==3.2.1
pypdfium2==4.30.1
pytest-asyncio==0.25.3
pytest-benchmark==5.1.0
pytest-cov==6.0.0
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-magic==0.4.27
pytz==2025.1
qdrant-client==1.13.2
qwen-vl-utils==0.0.8
ranx==0.3.20
ray==2.42.1
referencing==0.36.2
regex==2024.11.6
rpds-py==0.22.3
rsa==4.9
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
sentencepiece==0.2.0
shapely==2.0.7
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.6
soxr==0.5.0.post1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
speechbrain==1.0.2
Sphinx==7.4.7
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
srsly==2.5.1
sympy==1.13.1
tabulate==0.9.0
thinc==8.3.4
threadpoolctl==3.5.0
tifffile==2024.8.30
tiktoken==0.9.0
tomli==2.2.1
trec-car-tools==2.6
typer==0.12.5
typing_extensions==4.12.2
tzdata==2025.1
ultralytics-thop==2.0.14
umap-learn==0.5.7
unlzw3==0.2.3
uritemplate==4.1.1
urllib3==2.3.0
validators==0.34.0
virtualenv==20.29.2
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
whisper-openai==1.0.0
word2number==1.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
zlib-state==0.1.9
greenlet==3.1.1
semantic-version==2.10.0
setuptools==75.8.0
setuptools-rust==1.10.2
docling==2.25.0
